requirements.txt
streamlit
langchain-openai
streamlit-mic-recorder
python-dotenv